pull-all: stash and restore uncommmitted changes
If the user calls this script with dirty tree, his work
will be stashed and (hopefully) reapplied when the script
finishes successfully. If the script aborts, there might
be some cleanup required (e.g. to fix a failed rebase)
so in that case the stash isn't automatically applied.

There might be some funny situations in which this won't
work properly, so it's better to pull with clean tree.