lily building: grab the commit to be compiled from LILYPOND_GIT
until now a commit was checked out in $source; if $source was outside
main repository (i.e. building in a clone repo) this meant that a
different commit could be compiled than is checkedout in LILYGIT. For
example, HEAD in LILYGIT and HEAD in cloned repo might be completely
different, and until now the script would compile HEAD of clone,
whihc is usually now what we want.

UPDATE: the downside is that i cannot use the script wit a dirty
LILYGIT tree, because the checkout doesn't work (aborts).  I need to
find a workaround.