build-lily: change how commits-to-be-built are passed to child repo
Until now, things were messy.  Running the script had a
chance of messing with $LILYPOND_GIT (in particular,
changing its HEAD) and it was impossible to compile an
uncommitted state, or a commit that wasn't reachable from
any of the branches (such commits aren't transferred between
repositories during a fetch - the child repo doesn't know
about some commits exising in the main repo).

Currently the state of the source code that should be
committed is passed from the main repository ($LILYPOND_GIT)
to the child repositories (residing in $LILYPOND_BUILD_DIR)
using a tagged commit. This allows to pass "detached"
commits, since the tag points at them and will be fetched
(when --tags flag in git fetch is used). Similarly, an
uncommitted source state is temporarily committed, the
commit is marked with the tag and fetched, but then the
original branch is restored to the previous, uncommitted
state.