build-lily: abort merge if it fails
Since the merge is done 'under the hood' (i.e. in detached HEAD),
we don't want to investigate it - just throw bad merge away so
that it doesn't bother us.  Keep in mind that we take care to
stash any uncommitted changes earlier, so nothing should get
intermingled with the merge.